requirements.txt
numpy==1.17.2
pandas==0.25.1
scikit-learn==0.22.1
tensorflow==1.15.2
optuna==0.17.0
mlflow==1.7.0
pyyaml==5.1.2
